pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "growlancer-sdr"
version = "0.1.0"
description = "Growlancer Automated SDR System"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.39.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "pyairtable>=2.3.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "tenacity>=8.2.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
]

[project.scripts]
sdr = "sdr.main:main"

[tool.setuptools.packages.find]
include = ["sdr*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
